+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 [^format]
 
 <!--
-Bug empty line in code block is still compressed
+BUG empty line in code block is still compressed
 todo todo CLI to import/export w/ OpenWebUI
 -->